Additional fixes for data decode
